Moved data to refreshData
